refactor: CodeViewer와 CodeFold를 Code 도메인으로 이동
