+++ src/app/ui/JotaiDevTools/JotaiDevTools.tsx
@@ -8,12 +8,12 @@ import * as themeAtoms from '@/entities/AppTheme/atoms';
 import * as appAtoms from '@/entities/AppView/model/atoms';
 import { store } from '@/entities/AppView/model/store';
 import * as canvasAtoms from '@/features/Canvas/model/atoms';
+import * as codeFoldAtoms from '@/features/Code/CodeFold/model/atoms';
 import * as focusModeAtoms from '@/features/Code/FocusMode/model/atoms';
 import * as explorerAtoms from '@/features/Explorer/model/atoms';
 import * as navigationAtoms from '@/features/File/Navigation/model/atoms';
 import * as filesAtoms from '@/features/File/OpenFiles/model/atoms';
 import * as searchAtoms from '@/features/Search/UnifiedSearch/model/atoms';
-import * as codeFoldAtoms from '@/pages/shared/features/CodeFold/model/atoms';
 import * as deadCodeAtoms from '@/pages/shared/features/DeadCode/model/atoms';
 import { JotaiDevToolsOpenButton } from './JotaiDevToolsOpenButton';
 import { type AtomUpdate, JotaiDevToolsPanel } from './JotaiDevToolsPanel';
