+++ src/features/Code/CodeFold/lib/collectFoldMetadata.ts
