+++ src/features/Code/CodeFold/lib/foldUtils.ts
