+++ src/features/Code/CodeViewer/core/renderer/lib/deadCodeHelpers.ts
