+++ src/features/Code/CodeViewer/core/renderer/lib/segmentBuilders.ts
