+++ src/features/Code/CodeViewer/core/renderer/segmentUtils.ts
