+++ src/features/Code/CodeViewer/model/segment.ts
