+++ src/features/Code/CodeViewer/ui/CodeLineSegment.tsx
