+++ src/features/Code/CodeViewer/ui/SegmentRenderer.tsx
@@ -8,10 +8,10 @@ import { useMemo } from 'react';
 import { useEditorTheme } from '@/entities/AppTheme/EditorThemeProvider';
 import type { CanvasNode } from '@/entities/CanvasNode/model/types';
 import { visibleNodeIdsAtom } from '@/features/Canvas/model/atoms';
+import { buildSegmentStyle } from '@/features/Code/CodeViewer/core/styler/styleBuilder';
 import type { CodeLine, CodeSegment, SegmentKind } from '@/features/Code/CodeViewer/core/types/codeLine';
+import { LocalVariableSegment } from '@/features/Code/CodeViewer/ui/segments/LocalVariableSegment';
 import { activeLocalVariablesAtom } from '@/features/Code/FocusMode/model/atoms';
-import { buildSegmentStyle } from '@/pages/shared/features/CodeViewer/core/styler/styleBuilder';
-import { LocalVariableSegment } from '@/pages/shared/features/CodeViewer/ui/segments/LocalVariableSegment';
 import { DependencyTokenSegment } from './segments/DependencyTokenSegment';
 import { ExpandSegment } from './segments/ExpandSegment';
 import { ExternalSegment } from './segments/ExternalSegment';
