+++ src/features/Code/CodeViewer/ui/TemplateClickableSegment.tsx
