+++ src/features/Code/CodeViewer/ui/VueTemplateSection.tsx
