+++ src/pages/slides/widgets/SlideDeck/ui/SlideContent.tsx
@@ -1,8 +1,8 @@
 import { useAtomValue, useSetAtom } from 'jotai';
 import { useEffect, useMemo } from 'react';
 import { filesAtom, graphDataAtom } from '@/entities/AppView/model/atoms';
-import CodeViewer from '@/pages/shared/features/CodeViewer/CodeViewer';
-import { renderCodeLinesDirect } from '@/pages/shared/features/CodeViewer/core/renderer/renderCodeLinesDirect';
+import CodeViewer from '@/features/Code/CodeViewer/CodeViewer';
+import { renderCodeLinesDirect } from '@/features/Code/CodeViewer/core/renderer/renderCodeLinesDirect';
 import {
   extractSlideChunks,
   getCurrentSlideChunk,
